Update rust.vim. Switch submodule remote, which can't possibly work.
The official plugin code moved out of the language codebase and into
it's own repository, so I was able to move the remote from the mirror I
was using to the real one. Now that I think of it, the two were NEVER
connected, so this could fail dramatically on other machines.