Sort out neovim config out of the gate.
Homesick wants different files in a different place. It was easy enough
to manually symlink some stuff, but we have homesick for that. The only
twist was that neovim writes stuff to ~/.config, which is already the
home for a lot of config and can't be completely overwritten. But
Homesick has a trick for that, too, so we're good.